liquadate added

+++ programs/lending2/src/instructions/repay.rs
@@ -70,7 +70,7 @@ pub fn process_repay(ctx: Context<Repay>, amount: u64) -> Result<()> {
 
   let bank = &mut ctx.accounts.bank;
 
-  bank.total_borrowed -= (bank.total_borrowed as f64 * E.powf(bank.interest_rate as f64 * time_diff)) as u64;
+  bank.total_borrowed -= (bank.total_borrowed as f64 * E.powf(bank.interest_rate as f64 * time_diff as f64)) as u64;
 
   let value_per_share = bank.total_borrowed as f64 / bank.total_borrowed_shares as f64;
